procstats: Remove the eq method from CPUTime
The `eq` method was unexported and as such was only used in tests. Given
that we exposed other operator-like methods it was a bit of an anomaly
in being unexported.

To make it doubly weird, the default-generated `==` operator does
exactly what one would expect for the concrete type, so there's *no*
need for an `eq` method. Just rip it out and udpate the 3 tests that
used it.