Make pparser.LineKVFileParser generic and typesafe
Improve type-safety by pulling in the recent changes making
LineKVFileParser generic. This guarantees that the index that has been
generated for one type cannot be used with another.

This also lets us eliminate a runtime check that was verifying that the
correct type was passed. (since that's now enforced at compile-time)